app: refactor nft & web3 controllers
NftController is not a simple controller that does not need contract client and therefore can be initiated on app load. Methods that required contract client (minting) are now inside web3 controller.
Web3 controller is now the class that takes care of EthereumController, Signer and ContractClient initiation and interaction with ClipIt contract