hardhat: add support for minters to mint token for creators
Creators address will be verified and signed by ClipIt API and this signature and creator address can be used in  function.

This removes the need for the  function. Market now stores minters address and pay outs 1% of first token price to minter as an incentive for viewers(minters) to mint clips for streamers(creators)